adiciona entidades, repositórios e configurações iniciais para o projeto

+++ backend/api-gestao-albuns/src/main/resources/db/migration/V1__create_tables.sql
@@ -1,13 +1,13 @@
 -- Tabela de Artistas
 CREATE TABLE artista (
-                         id SERIAL PRIMARY KEY,
+                         id BIGSERIAL PRIMARY KEY,
                          nome VARCHAR(255) NOT NULL,
                          created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 -- Tabela de Álbuns
 CREATE TABLE album (
-                       id SERIAL PRIMARY KEY,
+                       id BIGSERIAL PRIMARY KEY,
                        titulo VARCHAR(255) NOT NULL,
                        capa_url VARCHAR(500),
                        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
@@ -16,8 +16,8 @@ CREATE TABLE album (
 -- Tabela de Associação (Relacionamento N:N)
 -- Um artista tem vários álbuns, um álbum pode ter vários artistas (feats)
 CREATE TABLE artista_album (
-                               artista_id INTEGER NOT NULL,
-                               album_id INTEGER NOT NULL,
+                               artista_id BIGINT NOT NULL,
+                               album_id BIGINT NOT NULL,
                                PRIMARY KEY (artista_id, album_id),
                                CONSTRAINT fk_artista FOREIGN KEY (artista_id) REFERENCES artista(id),
                                CONSTRAINT fk_album FOREIGN KEY (album_id) REFERENCES album(id)
